requirements.txt
google-adk>=0.0.2
google-genai>=1.5.0
pydantic>=2.10.6
python-dotenv>=1.0.0
pytest>=7.3.1
absl-py>=2.2.1
google-cloud-firestore>=2.15.0
google-cloud-bigquery>=3.12.0
google-cloud-storage>=2.14.0
fastapi>=0.110.0
uvicorn>=0.25.0
python-multipart>=0.0.9 